add visitor count & fix new plants
