+++ src/pages/index.tsx
@@ -13,7 +13,10 @@ import dynamic from "next/dynamic";
 import Guide from "./guide";
 import Link from "next/link";
 import { Marquee } from "../components/Marquee";
-import { VisitorCount } from "../components/VisitorCount";
+import {
+  ActiveVisitorCount,
+  ClickCount,
+} from "../components/ActiveVisitorCount";
 import type { PlayProvider as PlayProviderType } from "@playhtml/react";
 const PlayProvider = dynamic(
   () => import("@playhtml/react").then((c) => c.PlayProvider),
@@ -22,6 +25,10 @@ const PlayProvider = dynamic(
 
 const Garden = dynamic(() => import("../components/Garden"), { ssr: false });
 
+export const SeedDate = new Date("2023-03-15");
+const GardenAge = Math.floor(
+  (new Date().getTime() - SeedDate.getTime()) / 1000 / 60 / 60 / 24
+);
 export const StartDate = new Date("2023-12-21");
 export const GardenGrowingDays = Math.floor(
   (new Date().getTime() - StartDate.getTime()) / 1000 / 60 / 60 / 24
@@ -56,11 +63,6 @@ export default function Home() {
   const hasClickedRef = useRef(false);
 
   const [hasLoaded, setHasLoaded] = useState(false);
-  // const [plantsIdentified, setPlantsIdentified] = useGlobalState(
-  //   "plantsIdentified",
-  //   0
-  // );
-  const plantsIdentified = 0;
 
   useEffect(() => {
     const startSoundHandler = () => {
@@ -150,7 +152,10 @@ export default function Home() {
         separator={" 🌑 "}
       ></Marquee>
       <PlayProvider>
-        <VisitorCount />
+        <div id="stats">
+          <ClickCount />
+          <ActiveVisitorCount />
+        </div>
       </PlayProvider>
       <hgroup>
         <h1>html garden</h1>
@@ -163,11 +168,9 @@ export default function Home() {
           {currentSpecies.length > 1 ? "are" : "is"} blooming.
         </p>
         <p>
-          we are in {seasonName}.{" "}
-          {hasLoaded
-            ? `the garden has been growing for
-          ${GardenGrowingDays} days. ${plantsIdentified} plants identified.`
-            : null}
+          We are in {seasonName}. The garden has been growing for{" "}
+          {GardenGrowingDays} days this season and was seeded {GardenAge} days
+          ago.
         </p>
         <p>
           enable sound?{" "}
